Go back function
The function emy.goBack() doesn’t work.

L 220 : showView wait for an element view, not an id.

The first navStack.pop() contains the current view. The old function
does a showView on the… current view so ; and that’s not what we
expect.

I test if the navStack.pop return the currentView. If its the current
view, i do a second navStack.pop() and it returns the right previous
view.